Improve the api about basecourse.

+++ back-end/controller/basecourse_api.go
@@ -44,18 +44,24 @@ func ReadAllBaseCourse_Handler(c *gin.Context) {
 // @Param		base		path	string	true	"Base Course Name"
 // @Success 200 "success"
 // @Failure 500 {object} response.DBesponse{error=response.MySQLCreateError} "mysql error"
+// @Failure 400 "not valid"
 // @Security ApiKeyAuth
 // @Router /basecourse/{base} [post]
 func CreateBaseCourse_Handler(c *gin.Context) {
 	jwt.Check_Baselevel(c)
 
 	name := c.Param("base")
-	flag := dao.CreateBaseCourse(name)
 
-	if flag {
-		response.SuccessResponse(c, name)
+	if !dao.ValidBaseCourse(name) {
+		flag := dao.CreateBaseCourse(name)
+
+		if flag {
+			response.SuccessResponse(c, name)
+		} else {
+			response.ErrMySQLCreateResponse(c, BaseCourse_Model)
+		}
 	} else {
-		response.ErrMySQLCreateResponse(c, BaseCourse_Model)
+		response.ErrBasecourseNotValidResponse(c)
 	}
 }
 
@@ -81,7 +87,6 @@ func UpdateBaseCourse_Handler(c *gin.Context) {
 	new_name := c.Param("new")
 
 	if dao.ValidBaseCourse(name) {
-
 		flag, err := dao.UpdateBaseCourse(name, new_name)
 
 		if err != nil {
@@ -109,22 +114,28 @@ func UpdateBaseCourse_Handler(c *gin.Context) {
 // @Success 204
 // @Failure 500 {object} response.DBesponse{error=response.MySQLDeleteError} "mysql error"
 // @Failure 400 "not valid"
+// @Failure 404 "not exists"
 // @Security ApiKeyAuth
 // @Router /basecourse/{base} [delete]
 func DeleteBaseCourse_Handler(c *gin.Context) {
 	jwt.Check_Baselevel(c)
 
 	name := c.Param("base")
-	flag, err := dao.DeleteBaseCourse(name)
 
-	if err != nil {
-		response.ErrMySQLDeleteResponse(c, BaseCourse_Model)
-	} else {
-		if flag {
-			response.NonContentResponse(c)
+	if dao.ValidBaseCourse(name) {
+		flag, err := dao.DeleteBaseCourse(name)
+
+		if err != nil {
+			response.ErrMySQLDeleteResponse(c, BaseCourse_Model)
 		} else {
-			response.ErrBasecourseNotValidResponse(c)
+			if flag {
+				response.NonContentResponse(c)
+			} else {
+				response.ErrBasecourseNotValidResponse(c)
+			}
 		}
+	} else {
+		response.ErrBasecourseNotExistsResponse(c)
 	}
 }
 
@@ -138,18 +149,24 @@ func DeleteBaseCourse_Handler(c *gin.Context) {
 // @Param		course_name			path	string	true	"Course Name"
 // @Success 200 "success"
 // @Failure 500 {object} response.DBesponse{error=response.MySQLReadAllError} "mysql error"
+// @Failure 404 "not exists"
 // @Security ApiKeyAuth
 // @Router /{course_name}/base [get]
 func ReadBaseCourseRelation_Handler(c *gin.Context) {
 	jwt.Check_authlevel_Instructor(c)
 	coursename := course.GetCourse(c)
 
-	base_course, flag := dao.ReadBaseCourseRelation(coursename)
-	if flag {
-		response.SuccessResponse(c, base_course)
+	if dao.ValidBaseCourseRelation(coursename) {
+		base_course, flag := dao.ReadBaseCourseRelation(coursename)
+		if flag {
+			response.SuccessResponse(c, base_course)
+		} else {
+			response.ErrMySQLReadResponse(c, BaseCourseRelation_Model)
+		}
 	} else {
-		response.ErrMySQLReadResponse(c, BaseCourseRelation_Model)
+		response.ErrBasecourseRelationNotExistsResponse(c)
 	}
+
 }
 
 // CreateBaseCourseRelation_Handler godoc
@@ -163,6 +180,7 @@ func ReadBaseCourseRelation_Handler(c *gin.Context) {
 // @Param		base				path	string	true	"Base Course Name"
 // @Success 200 "success"
 // @Failure 500 {object} response.DBesponse{error=response.MySQLCreateError} "mysql error"
+// @Failure 400 "not valid"
 // @Security ApiKeyAuth
 // @Router /{course_name}/{base} [post]
 func CreateBaseCourseRelation_Handler(c *gin.Context) {
@@ -170,12 +188,17 @@ func CreateBaseCourseRelation_Handler(c *gin.Context) {
 
 	base := c.Param("base")
 	coursename := course.GetCourse(c)
-	flag := dao.CreateBaseCourseRelation(coursename, base)
 
-	if flag {
-		response.SuccessResponse(c, coursename+" "+base)
+	if !dao.ValidBaseCourseRelation(coursename) {
+		flag := dao.CreateBaseCourseRelation(coursename, base)
+
+		if flag {
+			response.SuccessResponse(c, coursename+" "+base)
+		} else {
+			response.ErrMySQLCreateResponse(c, BaseCourseRelation_Model)
+		}
 	} else {
-		response.ErrMySQLCreateResponse(c, BaseCourseRelation_Model)
+		response.ErrBasecourseRelationRecreatedResponse(c)
 	}
 }
 
@@ -190,6 +213,7 @@ func CreateBaseCourseRelation_Handler(c *gin.Context) {
 // @Param		base				path	string	true	"Base Course Name"
 // @Success 200 "success"
 // @Failure 500 {object} response.DBesponse{error=response.MySQLUpdateError} "mysql error"
+// @Failure 404 "not exists"
 // @Security ApiKeyAuth
 // @Router /{course_name}/{base} [put]
 func UpdateBaseCourseRelation_Handler(c *gin.Context) {
@@ -198,13 +222,22 @@ func UpdateBaseCourseRelation_Handler(c *gin.Context) {
 	base := c.Param("base")
 	coursename := course.GetCourse(c)
 
-	err := dao.UpdateBaseCourseRelation(coursename, base)
+	if dao.ValidBaseCourseRelation(coursename) {
+		if flag, _ := dao.ValidateAssessmentByCourse(coursename); flag {
+			err := dao.UpdateBaseCourseRelation(coursename, base)
 
-	if err != nil {
-		response.ErrMySQLUpdateResponse(c, BaseCourseRelation_Model)
+			if err != nil {
+				response.ErrMySQLUpdateResponse(c, BaseCourseRelation_Model)
+			} else {
+				response.SuccessResponse(c, coursename+" "+base)
+			}
+		} else {
+			response.ErrBasecourseRelationNotValidResponse(c)
+		}
 	} else {
-		response.SuccessResponse(c, coursename+" "+base)
+		response.ErrBasecourseRelationNotExistsResponse(c)
 	}
+
 }
 
 // DeleteBaseCourseRelation_Handler godoc
@@ -217,6 +250,7 @@ func UpdateBaseCourseRelation_Handler(c *gin.Context) {
 // @Param		course_name			path	string	true	"Course Name"
 // @Success 204
 // @Failure 500 {object} response.DBesponse{error=response.MySQLDeleteError} "mysql error"
+// @Failure 404 "not exists"
 // @Security ApiKeyAuth
 // @Router /{course_name}/base [delete]
 func DeleteBaseCourseRelation_Handler(c *gin.Context) {
@@ -224,11 +258,19 @@ func DeleteBaseCourseRelation_Handler(c *gin.Context) {
 
 	coursename := course.GetCourse(c)
 
-	err := dao.DeleteBaseCourseRelation(coursename)
+	if dao.ValidBaseCourseRelation(coursename) {
+		if flag, _ := dao.ValidateAssessmentByCourse(coursename); flag {
+			err := dao.DeleteBaseCourseRelation(coursename)
 
-	if err != nil {
-		response.ErrMySQLDeleteResponse(c, BaseCourse_Model)
+			if err != nil {
+				response.ErrMySQLDeleteResponse(c, BaseCourse_Model)
+			} else {
+				response.NonContentResponse(c)
+			}
+		} else {
+			response.ErrBasecourseRelationNotValidResponse(c)
+		}
 	} else {
-		response.NonContentResponse(c)
+		response.ErrBasecourseRelationNotExistsResponse(c)
 	}
 }
